extract_body_size_range_match()

calculate whether target length matches equation range

Usage

extract_body_size_range_match(
  equation_id,
  target_body_size,
  prop = 0.3,
  equation_db
)

Arguments

equation_id

id number of the equation

target_body_size

body size of the target taxon

prop

proportion above and below equation body size range considered acceptable for using an equation

equation_db

equation database object

Value

string with the genus name

Details

This function generates a data.frame with three variables: body_size_range_match: whether the target length is within the equation length range (TRUE or FALSE), body_size_min_dist: percentage away from the minimum length used to create the equation, body_size_max_dist: percentage away from the maximum length.

Author

James G. Hagan (james_hagan(at)outlook.com)