get_habitat_data()

Get habitat data from Abell et al.'s (2008) freshwater ecoregion map

Usage

get_habitat_data(data, latitude_dd, longitude_dd)

Arguments

data

data.frame with a column containing latitude and longitude in decimal degrees

latitude_dd

character string with the column name containing the latitude in decimal degrees

longitude_dd

character string with the column name containing the longitude in decimal degrees

Value

tibble of the input data with habitat data from Abell et al.s (2008) ecoregion map attached as additional columns function to get habitat data

Details

This function takes a data.frame with latitude and longitude data and gets data on biogeographic realm, major habitat type and ecoregion based on Abell et al.'s (2008) freshwater ecoregion map.

Author

James G. Hagan (james_hagan(at)outlook.com) and Ronald Bergmann