Invertebrates collected in temporary ponds

The test data consist of invertebrate taxa collected in six temporary ponds in a nature reserve in Belgium. In addition, the data consist of latitude-longitude coordinates for the samples as well as measured body lengths for all the taxa.

Usage

data(ponds)

Format

An object of class data.frame

References

Vincent Dolmans (Vincent.Dolmans@vub.be)

Examples

data(ponds)