select_traits_tax_dist()

Get taxonomic distances of target names relative to the taxa databases

Usage

select_traits_tax_dist(
  data,
  target_taxon,
  body_size,
  life_stage,
  max_tax_dist = 3,
  trait = "equation",
  gen_sp_dist = 0.5
)

Arguments

data

input data.frame exported from get_habitat_data() and clean_taxon_names() function

target_taxon

character string with the column name containing the taxon names

body_size

column name containing the body length data

life_stage

character string with the column name containing the life-stage information

max_tax_dist

maximum taxonomic distance acceptable between the target and the taxa in the database (default = 3)

trait

trait to be searched for (default = "equation")

gen_sp_dist

taxonomic distance between a genus and a species (default = 0.5)

Value

tibble of the input data with traits or equations within the maximum taxonomic distance

Details

This function searches the relevant trait or equation database for the best matching trait or equation for a given target name based on taxonomic distance

Author

James G. Hagan (james_hagan(at)outlook.com)